Move PluginNamespace array into PluginManager
